+++ calendar-service/package.json
@@ -38,14 +38,13 @@
     "@nestjs/serve-static": "^5.0.4",
     "@nestjs/swagger": "^11.2.0",
     "@nestjs/websockets": "^11.1.9",
-    "@types/pdfkit": "^0.17.5",
     "axios": "^1.13.1",
     "cookie-parser": "^1.4.7",
     "drizzle-orm": "^0.44.7",
     "exceljs": "^4.4.0",
     "nestjs-zod": "^5.0.0",
-    "pdfkit": "^0.18.0",
     "postgres": "^3.4.3",
+    "puppeteer": "^24.40.0",
     "reflect-metadata": "^0.2.2",
     "rxjs": "^7.8.1",
     "socket.io": "^4.8.1",
