+++ calendar-service/src/reports/pdf-generator.service.ts
@@ -1,0 +1,26 @@
+import { Injectable } from '@nestjs/common';
+import puppeteer from 'puppeteer';
+
+@Injectable()
+export class PdfGeneratorService {
+  async generatePdfFromHtml(html: string): Promise<Buffer> {
+    const browser = await puppeteer.launch({
+      headless: true,
+      args: ['--no-sandbox', '--disable-setuid-sandbox'],
+    });
+
+    try {
+      const page = await browser.newPage();
+      await page.setContent(html, {
+        waitUntil: 'networkidle0',
+      });
+      const pdf = await page.pdf({
+        format: 'A4',
+        printBackground: true,
+      });
+      return Buffer.from(pdf);
+    } finally {
+      await browser.close();
+    }
+  }
+}
