+++ calendar-service/src/reports/reports.module.ts
@@ -2,13 +2,14 @@ import { Module } from '@nestjs/common';
 
 import { ActivitiesModule } from '../activities/activities.module';
 import { DatabaseModule } from '../database/database.module';
+import { PdfGeneratorService } from './pdf-generator.service';
 import { ReportsController } from './reports.controller';
 import { ReportsService } from './reports.service';
 
 @Module({
   imports: [DatabaseModule, ActivitiesModule],
   controllers: [ReportsController],
-  providers: [ReportsService],
+  providers: [ReportsService, PdfGeneratorService],
   exports: [ReportsService],
 })
 export class ReportsModule {}
