+++ packages/shared/package.json
@@ -71,6 +71,12 @@
       "import": "./dist/esm/reports/reportExportFormat.js",
       "require": "./dist/cjs/reports/reportExportFormat.js",
       "default": "./dist/esm/reports/reportExportFormat.js"
+    },
+    "./reports/reportPrintHtml": {
+      "types": "./dist/esm/reports/reportPrintHtml.d.ts",
+      "import": "./dist/esm/reports/reportPrintHtml.js",
+      "require": "./dist/cjs/reports/reportPrintHtml.js",
+      "default": "./dist/esm/reports/reportPrintHtml.js"
     }
   },
   "scripts": {
