+++ packages/shared/src/reports/reportPrintHtml.ts
@@ -1,0 +1,6 @@
+export {
+  getReportTemplateHtml,
+  wrapReportHtmlDocument,
+} from './getReportTemplateHtml';
+export { buildExecLookAheadPrintHTML } from './print/execLookAheadPrintHTML';
+export { buildLookAheadLegacyPrintHtml } from './print/lookAheadLegacyPrintHtml';
